+++ src/tg2go/db/models/order.py
@@ -1,0 +1,68 @@
+from datetime import datetime
+from enum import Enum
+from uuid import UUID, uuid4
+
+from sqlalchemy import DateTime, Enum as SqlEnum, ForeignKey, Numeric, Text, text
+from sqlalchemy.orm import Mapped, mapped_column, relationship
+
+from tg2go.db.base import Base
+from tg2go.db.models.order_item import OrderItem
+
+
+# TODO: add more order statuses
+class OrderStatus(str, Enum):
+    pending = "pending"
+    paid = "paid"
+    in_progress = "in_progress"
+    done = "done"
+    cancelled = "cancelled"
+
+
+class Order(Base):
+    __tablename__ = "orders"
+
+    # --- primary key ---
+    order_id: Mapped[UUID] = mapped_column(
+        primary_key=True,
+        default=uuid4,
+        nullable=False,
+    )
+
+    # --- secondary key ---
+    chat_id: Mapped[int] = mapped_column(
+        ForeignKey("users.chat_id"),
+        index=True,
+    )
+
+    # --- description ---
+    status: Mapped[OrderStatus] = mapped_column(
+        SqlEnum(OrderStatus),
+        nullable=False,
+    )
+    total_price_rub: Mapped[float] = mapped_column(
+        Numeric(10, 2),
+    )
+    internal_comment: Mapped[str] = mapped_column(
+        Text,
+        nullable=True,
+    )
+    client_comment: Mapped[str] = mapped_column(
+        Text,
+        nullable=True,
+    )
+
+    # --- time ---
+    created_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True),
+        server_default=text("CURRENT_TIMESTAMP"),
+        nullable=False,
+    )
+    updated_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True),
+        server_default=text("CURRENT_TIMESTAMP"),
+        onupdate=text("CURRENT_TIMESTAMP"),
+        nullable=False,
+    )
+
+    # --- relationship ---
+    order_items: Mapped[list[OrderItem]] = relationship(back_populates="order")